refactor(overpass): one client, one cache, no cassettes
Unify all Overpass API usage into a single TypeScript client
(scripts/pipeline/lib/overpass.ts) with:
- Content-addressed per-query cache in .cache/overpass/
- Built-in concurrency semaphore (max 4 parallel fetches)
- In-flight dedup, server rotation, auto-recording

Remove the cassette system (createRecorder, RECORD_OVERPASS env var),
the old .mjs client, and the unused overpass-pool.ts. CI restores
.cache/overpass/ once before prebuild — no separate recording step.
