+++ .github/workflows/_build-city.yml
@@ -151,13 +151,13 @@ jobs:
         if: inputs.deploy
         run: echo "BUILD_START=$(date -u +%Y-%m-%dT%H:%M:%SZ)" >> "$GITHUB_ENV"
 
-      - name: Restore pipeline cassette cache
+      - name: Restore Overpass cache
         if: ${{ !inputs.deploy-only }}
         uses: actions/cache@v5
         with:
-          path: scripts/pipeline/.cache
-          key: pipeline-cassette-${{ hashFiles('scripts/pipeline/build-bikepaths.ts', 'scripts/pipeline/lib/overpass.mjs') }}
-          restore-keys: pipeline-cassette-
+          path: .cache/overpass
+          key: overpass-${{ hashFiles('scripts/pipeline/lib/overpass.ts') }}
+          restore-keys: overpass-
 
       - name: Run prebuild
         if: ${{ !inputs.deploy-only }}
