+++ .github/workflows/_test.yml
@@ -69,13 +69,12 @@ jobs:
         env:
           PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD: '1'
 
-      - name: Restore pipeline cassette cache
-        id: cassette-cache
+      - name: Restore Overpass cache
         uses: actions/cache@v5
         with:
-          path: scripts/pipeline/.cache
-          key: pipeline-cassette-${{ hashFiles('scripts/pipeline/build-bikepaths.ts', 'scripts/pipeline/lib/overpass.mjs') }}
-          restore-keys: pipeline-cassette-
+          path: .cache/overpass
+          key: overpass-${{ hashFiles('scripts/pipeline/lib/overpass.ts') }}
+          restore-keys: overpass-
 
       - name: Run prebuild
         run: npx tsx scripts/prebuild.ts
@@ -83,12 +82,6 @@ jobs:
           CONTENT_DIR: ${{ github.workspace }}/bike-routes
           CITY: demo
           PREBUILD_MINIMAL: '1'
-
-      - name: Record pipeline cassettes
-        if: steps.cassette-cache.outputs.cache-hit != 'true'
-        run: RECORD_OVERPASS=ottawa node scripts/pipeline/build-bikepaths.ts --city ottawa
-        env:
-          CONTENT_DIR: ${{ github.workspace }}/bike-routes
 
       - name: Run unit tests
         run: npx vitest run
@@ -216,12 +209,12 @@ jobs:
         env:
           PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD: '1'
 
-      - name: Restore pipeline cassette cache
+      - name: Restore Overpass cache
         uses: actions/cache@v5
         with:
-          path: scripts/pipeline/.cache
-          key: pipeline-cassette-${{ hashFiles('scripts/pipeline/build-bikepaths.ts', 'scripts/pipeline/lib/overpass.mjs') }}
-          restore-keys: pipeline-cassette-
+          path: .cache/overpass
+          key: overpass-${{ hashFiles('scripts/pipeline/lib/overpass.ts') }}
+          restore-keys: overpass-
 
       - name: Run prebuild
         run: npx tsx scripts/prebuild.ts
