--- scripts/pipeline/lib/overpass.mjs
@@ -1,599 +1,0 @@
-/**
- * Overpass API client with file-based caching.
- */
-
-import { createHash } from 'node:crypto';
-import { mkdirSync, existsSync, readFileSync, writeFileSync } from 'node:fs';
-import { join, dirname } from 'node:path';
-import { fileURLToPath } from 'node:url';
-
-const __dirname = dirname(fileURLToPath(import.meta.url));
-const CACHE_DIR = join(__dirname, '..', '.cache');
-
-// Server rotation — try our own server first, then public fallbacks.
-const OVERPASS_SERVERS = [
-  'https://overpass.whereto.bike/api/interpreter',
-  'https://overpass.private.coffee/api/interpreter',
-  'https://overpass-api.de/api/interpreter',
-];
-const OVERPASS_URL = process.env.OVERPASS_URL || OVERPASS_SERVERS[0];
-
-function cacheKey(query) {
-  const hash = createHash('sha256').update(query).digest('base64url');
-  return hash.slice(0, 40);
-}
-
-function normQ(q) {
-  return q.replace(/\s+/g, ' ').trim();
-}
-
-/**
- * Content-addressed cassette format (Nix-like append-only cache).
- *
- * Each query is hashed. Responses are stored by hash. A cached response
- * for a given hash is valid as long as OSM data hasn't changed — code
- * changes don't invalidate existing entries.
- *
- * Recording is additive: existing entries are preserved, only missing
- * hashes are fetched from the API. Never delete the cassette to re-record.
- *
- * Three modes:
- * 1. RECORD_OVERPASS set → record mode. Load existing, fetch misses, save combined.
- * 2. Cassette exists, no env var → play mode. Return cached data. Throw on miss.
- * 3. No cassette, no env var → hard error (tests must not silently skip).
- */
-function loadCassette(name) {
-  const cassettePath = join(CACHE_DIR, `cassette-${name}.json`);
-  if (!existsSync(cassettePath)) return null;
-  const raw = JSON.parse(readFileSync(cassettePath, 'utf8'));
-  // Migrate: old format was [{query, data}], new format is {hash: {query, data}}
-  if (Array.isArray(raw)) {
-    const migrated = {};
-    for (const entry of raw) {
-      const key = cacheKey(normQ(entry.query));
-      migrated[key] = { query: entry.query, data: entry.data };
-    }
-    return migrated;
-  }
-  return raw;
-}
-
-function saveCassette(name, entries) {
-  const cassettePath = join(CACHE_DIR, `cassette-${name}.json`);
-  mkdirSync(CACHE_DIR, { recursive: true });
-  writeFileSync(cassettePath, JSON.stringify(entries));
-}
-
-/**
- * Create an additive recorder. Loads existing cassette entries (if any),
- * returns cached data for known hashes, fetches and appends for misses.
- *
- * @param {string} name cassette filename
- * @param {object} [opts]
- * @param {(query: string) => Promise<{elements: any[]}>} [opts.underlying]
- *   override the network fetcher (defaults to module-level queryOverpass)
- * @param {string} [opts.cacheDir] override the cassette directory
- *
- * Record new entries:
- *   RECORD_OVERPASS=ottawa node scripts/pipeline/build-bikepaths.mjs --city ottawa
- */
-export function createRecorder(name, opts = {}) {
-  const underlying = opts.underlying || queryOverpass;
-  const dir = opts.cacheDir || CACHE_DIR;
-  const cassettePath = join(dir, `cassette-${name}.json`);
-  const existing = loadCassetteFrom(cassettePath);
-  const entries = existing || {};
-  let newCount = 0;
-  // In-flight dedupe: parallel fetches for the same query share one Promise
-  const inFlight = new Map();
-  const recorder = async (query) => {
-    const key = cacheKey(normQ(query));
-    if (entries[key]) return entries[key].data;
-    if (inFlight.has(key)) return inFlight.get(key);
-    const promise = (async () => {
-      try {
-        const data = await underlying(query);
-        entries[key] = { query, data };
-        newCount++;
-        mkdirSync(dir, { recursive: true });
-        writeFileSync(cassettePath, JSON.stringify(entries));
-        return data;
-      } finally {
-        inFlight.delete(key);
-      }
-    })();
-    inFlight.set(key, promise);
-    return promise;
-  };
-  recorder.stats = () => ({ total: Object.keys(entries).length, new: newCount });
-  return recorder;
-}
-
-function loadCassetteFrom(cassettePath) {
-  if (!existsSync(cassettePath)) return null;
-  const raw = JSON.parse(readFileSync(cassettePath, 'utf8'));
-  if (Array.isArray(raw)) {
-    const migrated = {};
-    for (const entry of raw) {
-      const key = cacheKey(normQ(entry.query));
-      migrated[key] = { query: entry.query, data: entry.data };
-    }
-    return migrated;
-  }
-  return raw;
-}
-
-
-/**
- * In-flight query tracking. If a query is already being fetched,
- * callers share the same Promise instead of hitting the server twice.
- * This prevents the Overpass dispatcher's duplicate_query rejection
- * when parallel phases fire identical queries (e.g. the same adapter
- * query appearing twice in the namedWayQueries list).
- * @type {Map<string, Promise<{elements: any[]}>>}
- */
-const queryInFlight = new Map();
-
-/**
- * POST a query to Overpass, caching results to disk.
- * Returns parsed JSON response.
- */
-export async function queryOverpass(query) {
-  mkdirSync(CACHE_DIR, { recursive: true });
-
-  const key = cacheKey(query);
-  const cachePath = join(CACHE_DIR, `overpass-${key}.json`);
-
-  if (existsSync(cachePath)) {
-    console.log(`[overpass] cache hit: ${key}`);
-    return JSON.parse(readFileSync(cachePath, 'utf8'));
-  }
-
-  // Dedupe: if this exact query is already in flight, share the result
-  if (queryInFlight.has(key)) {
-    console.log(`[overpass] in-flight dedupe: ${key}`);
-    return queryInFlight.get(key);
-  }
-
-  console.log(`[overpass] fetching from API (key: ${key})`);
-
-  const promise = (async () => {
-    let data;
-    for (let attempt = 0; attempt < OVERPASS_SERVERS.length * 2; attempt++) {
-      const serverIdx = Math.floor(attempt / 2) % OVERPASS_SERVERS.length;
-      const serverUrl = process.env.OVERPASS_URL || OVERPASS_SERVERS[serverIdx];
-      const res = await fetch(serverUrl, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-        body: `data=${encodeURIComponent(query)}`,
-      });
-
-      if (res.ok) {
-        const contentType = res.headers.get('content-type') || '';
-        if (!contentType.includes('json')) {
-          // Server returned 200 but not JSON — read the body to diagnose
-          const htmlBody = await res.text();
-          if (htmlBody.includes('duplicate_query')) {
-            console.error(`\n[overpass] FATAL: duplicate_query error from ${serverUrl}`);
-            console.error(`[overpass] query key: ${key}`);
-            console.error(`[overpass] query (first 300 chars): ${query.replace(/\s+/g, ' ').slice(0, 300)}`);
-            console.error(`[overpass] This means two identical queries were sent in parallel.`);
-            console.error(`[overpass] Full error: ${htmlBody.replace(/\s+/g, ' ').slice(0, 200)}`);
-            process.exit(1);
-          }
-          console.log(`[overpass] ${serverUrl} returned 200 but content-type=${contentType}, trying next server...`);
-          continue;
-        }
-        const parsed = await res.json();
-        if (parsed.remark && parsed.remark.includes('runtime error')) {
-          console.log(`[overpass] ${serverUrl} returned runtime error: ${parsed.remark.slice(0, 80)}, trying next server...`);
-          continue;
-        }
-        data = parsed;
-        break;
-      }
-
-      if (res.status === 504 || res.status === 429 || res.status === 502 || res.status === 503) {
-        const wait = (attempt + 1) * 10;
-        const nextServer = OVERPASS_SERVERS[(serverIdx + 1) % OVERPASS_SERVERS.length];
-        console.log(`[overpass] ${serverUrl} returned ${res.status}, trying next server in ${wait}s...`);
-        await new Promise((r) => setTimeout(r, wait * 1000));
-        continue;
-      }
-
-      throw new Error(`Overpass API error ${res.status}: ${await res.text()}`);
-    }
-
-    if (!data) throw new Error('Overpass API: failed after 3 retries');
-    writeFileSync(cachePath, JSON.stringify(data), 'utf8');
-    console.log(`[overpass] cached ${data.elements?.length ?? 0} elements to ${key}`);
-    return data;
-  })();
-
-  queryInFlight.set(key, promise);
-  try {
-    return await promise;
-  } finally {
-    queryInFlight.delete(key);
-  }
-}
-
-/**
- * Fetch points of interest within a bounding box.
- * @param {[number, number, number, number]} bounds - [south, west, north, east]
- * @returns {Array<{ name, lat, lng, type, osmType, osmId, tags }>}
- */
-export async function fetchPOIs(bounds) {
-  const [s, w, n, e] = bounds;
-  const bbox = `${s},${w},${n},${e}`;
-
-  // POI types that matter for cycling: destinations worth riding to,
-  // places to stop, and cycling-specific infrastructure.
-  // Modeled after the curated places in Ottawa (128 hand-picked spots).
-  const query = `
-[out:json][timeout:180];
-(
-  // Parks, gardens, beaches — primary ride destinations
-  way["leisure"="park"](${bbox});
-  node["leisure"="park"](${bbox});
-  way["leisure"="garden"]["name"](${bbox});
-  node["natural"="beach"](${bbox});
-  way["natural"="beach"](${bbox});
-
-  // Viewpoints, squares, plazas — scenic stops
-  node["tourism"="viewpoint"](${bbox});
-  node["place"="square"](${bbox});
-  way["place"="square"](${bbox});
-
-  // Food & drink — where cyclists refuel
-  node["amenity"="cafe"]["name"](${bbox});
-  node["amenity"="ice_cream"]["name"](${bbox});
-  node["amenity"="pub"]["name"](${bbox});
-  node["amenity"="marketplace"](${bbox});
-  way["amenity"="marketplace"](${bbox});
-
-  // Cycling infrastructure
-  node["amenity"="bicycle_rental"](${bbox});
-  node["shop"="bicycle"](${bbox});
-
-  // Transport hubs
-  node["railway"="station"](${bbox});
-  node["amenity"="ferry_terminal"](${bbox});
-
-  // Camping (for longer rides)
-  node["tourism"="camp_site"]["name"](${bbox});
-  way["tourism"="camp_site"]["name"](${bbox});
-
-  // Culture
-  node["tourism"="museum"]["name"](${bbox});
-  way["tourism"="museum"]["name"](${bbox});
-  way["natural"="water"]["name"](${bbox});
-
-  // Bridges — river crossings make rides memorable
-  way["bridge"="yes"]["name"]["highway"~"cycleway|path|footway|pedestrian"](${bbox});
-  way["man_made"="bridge"]["name"](${bbox});
-);
-out center tags;
-`.trim();
-
-  const data = await queryOverpass(query);
-
-  const results = [];
-  for (const el of data.elements ?? []) {
-    // Extract lat/lng — ways have center, nodes have direct coords
-    let lat, lng;
-    if (el.center) {
-      lat = el.center.lat;
-      lng = el.center.lon;
-    } else if (el.lat != null) {
-      lat = el.lat;
-      lng = el.lon;
-    }
-
-    if (lat == null || lng == null) continue;
-
-    const tags = el.tags ?? {};
-    const name = tags.name;
-    if (!name) continue;
-
-    // Determine type from first matching tag
-    let type = null;
-    if (tags.bridge || tags.man_made === 'bridge') type = 'bridge';
-    else if (tags.leisure) type = tags.leisure;
-    else if (tags.tourism) type = tags.tourism;
-    else if (tags.amenity) type = tags.amenity;
-    else if (tags.place) type = tags.place;
-    else if (tags.railway) type = tags.railway;
-    else if (tags.natural) type = tags.natural;
-
-    if (!type) continue;
-
-    results.push({
-      name,
-      lat,
-      lng,
-      type,
-      osmType: el.type,
-      osmId: el.id,
-      tags,
-    });
-  }
-
-  return results;
-}
-
-/**
- * Fetch cycling infrastructure ways within a bounding box.
- * @param {[number, number, number, number]} bounds - [south, west, north, east]
- * @returns {Array} OSM way elements with geometry (length >= 2 nodes)
- */
-export async function fetchCyclingWays(bounds) {
-  const [s, w, n, e] = bounds;
-  const bbox = `${s},${w},${n},${e}`;
-
-  const query = `
-[out:json][timeout:60];
-(
-  way["highway"="cycleway"](${bbox});
-  way["cycleway"~"track|lane|shared_lane"](${bbox});
-  way["cycleway:left"~"track|lane"](${bbox});
-  way["cycleway:right"~"track|lane"](${bbox});
-  way["bicycle"="designated"]["highway"~"path|footway"](${bbox});
-);
-out geom tags;
-`.trim();
-
-  const data = await queryOverpass(query);
-
-  return (data.elements ?? []).filter(
-    (el) => Array.isArray(el.geometry) && el.geometry.length >= 2,
-  );
-}
-
-/**
- * Fetch metro/light-rail stations within a bounding box.
- * @param {[number, number, number, number]} bounds - [south, west, north, east]
- * @returns {Array<{ name, lat, lng, type, osmType, osmId, tags }>}
- */
-export async function fetchMetroStations(bounds) {
-  const [s, w, n, e] = bounds;
-  const bbox = `${s},${w},${n},${e}`;
-  const query = `
-[out:json][timeout:30];
-(
-  node["railway"="station"]["station"="subway"](${bbox});
-  node["railway"="station"]["station"="light_rail"](${bbox});
-);
-out tags;
-`.trim();
-
-  const data = await queryOverpass(query);
-  return (data.elements ?? [])
-    .filter((el) => el.tags?.name)
-    .map((el) => ({
-      name: el.tags.name,
-      lat: el.lat,
-      lng: el.lon,
-      type: 'metro',
-      osmType: el.type,
-      osmId: el.id,
-      tags: el.tags,
-    }));
-}
-
-/**
- * Fetch named rivers and canals within a bounding box.
- * @param {[number, number, number, number]} bounds - [south, west, north, east]
- * @returns {Array<{ name, geometry }>} geometry is array of [lon, lat] pairs
- */
-export async function fetchWaterways(bounds) {
-  const [s, w, n, e] = bounds;
-  const bbox = `${s},${w},${n},${e}`;
-  const query = `
-[out:json][timeout:30];
-(
-  way["waterway"="river"]["name"](${bbox});
-  way["waterway"="canal"]["name"](${bbox});
-);
-out geom tags;
-`.trim();
-
-  const data = await queryOverpass(query);
-  return (data.elements ?? [])
-    .filter((el) => Array.isArray(el.geometry) && el.geometry.length >= 2)
-    .map((el) => ({
-      name: el.tags.name,
-      geometry: el.geometry.map(({ lat, lon }) => [lon, lat]),
-    }));
-}
-
-/**
- * Fetch all rideable roads within a bounding box for gap routing.
- * @param {[number, number, number, number]} bounds - [south, west, north, east]
- * @returns {Array} OSM way elements with geometry
- */
-export async function fetchRoadNetwork(bounds) {
-  const [s, w, n, e] = bounds;
-  const bbox = `${s},${w},${n},${e}`;
-  const query = `
-[out:json][timeout:90];
-(
-  way["highway"~"cycleway|path|footway|residential|tertiary|secondary|living_street|pedestrian|service"](${bbox});
-);
-out geom;
-`.trim();
-
-  const data = await queryOverpass(query);
-  return (data.elements ?? [])
-    .filter(el => Array.isArray(el.geometry) && el.geometry.length >= 2);
-}
-
-/**
- * Fetch motorway/trunk/motorway_link ways within a bounding box.
- * @param {[number, number, number, number]} bounds - [south, west, north, east]
- * @returns {Array<Array<[number, number]>>} array of coordinate arrays ([lon, lat] pairs)
- */
-export async function fetchMotorways(bounds) {
-  const [s, w, n, e] = bounds;
-  const bbox = `${s},${w},${n},${e}`;
-  const query = `
-[out:json][timeout:30];
-(
-  way["highway"="motorway"](${bbox});
-  way["highway"="trunk"](${bbox});
-  way["highway"="motorway_link"](${bbox});
-);
-out geom;
-`.trim();
-
-  const data = await queryOverpass(query);
-  return (data.elements ?? [])
-    .filter((el) => Array.isArray(el.geometry) && el.geometry.length >= 2)
-    .map((el) => el.geometry.map(({ lat, lon }) => [lon, lat]));
-}
-
-/**
- * Fetch all "interesting things" for zone clustering — cast a wide net.
- * @param {[number, number, number, number]} bounds - [south, west, north, east]
- * @returns {Array<{ lat, lng, tags }>}
- */
-export async function fetchZonePOIs(bounds) {
-  const [s, w, n, e] = bounds;
-  const bbox = `${s},${w},${n},${e}`;
-  const query = `
-[out:json][timeout:120];
-(
-  node["amenity"~"cafe|restaurant|bar|pub|fast_food|ice_cream|food_court|marketplace|theatre|cinema|arts_centre|biergarten"](${bbox});
-  node["shop"~"bakery|deli|wine|chocolate|books"](${bbox});
-  node["tourism"~"museum|gallery|viewpoint|attraction|artwork"](${bbox});
-  node["leisure"~"beer_garden|playground|swimming_pool"](${bbox});
-  node["historic"~"monument|memorial|castle"](${bbox});
-);
-out;
-`.trim();
-
-  const data = await queryOverpass(query);
-  return (data.elements ?? [])
-    .filter((el) => el.lat != null && el.lon != null)
-    .map((el) => ({ lat: el.lat, lng: el.lon, tags: el.tags ?? {} }));
-}
-
-/**
- * Fetch tree rows for shade/canopy scoring.
- * @param {[number, number, number, number]} bounds
- * @returns {Array<Array<[number, number]>>} array of coordinate arrays [lon, lat]
- */
-export async function fetchTreeRows(bounds) {
-  const [s, w, n, e] = bounds;
-  const bbox = `${s},${w},${n},${e}`;
-  const query = `
-[out:json][timeout:30];
-way["natural"="tree_row"](${bbox});
-out geom;
-`.trim();
-
-  const data = await queryOverpass(query);
-  return (data.elements ?? [])
-    .filter((el) => Array.isArray(el.geometry) && el.geometry.length >= 2)
-    .map((el) => el.geometry.map(({ lat, lon }) => [lon, lat]));
-}
-
-/**
- * Fetch bike parking racks and rental stations.
- * @param {[number, number, number, number]} bounds
- * @returns {Array<{ lat, lng, type }>}
- */
-export async function fetchBikeParking(bounds) {
-  const [s, w, n, e] = bounds;
-  const bbox = `${s},${w},${n},${e}`;
-  const query = `
-[out:json][timeout:30];
-(
-  node["amenity"="bicycle_parking"](${bbox});
-  node["amenity"="bicycle_rental"](${bbox});
-);
-out;
-`.trim();
-
-  const data = await queryOverpass(query);
-  return (data.elements ?? [])
-    .filter((el) => el.lat != null)
-    .map((el) => ({ lat: el.lat, lng: el.lon, type: el.tags?.amenity }));
-}
-
-/**
- * Fetch park/plaza/garden areas with polygon geometry.
- * Ways get geometry directly; relations have member geometry.
- * @param {[number,number,number,number]} bounds - [south, west, north, east]
- * @returns {Array<{ name, type, lat, lng, geometry, extent, osmType, osmId, tags }>}
- */
-export async function fetchParkAreas(bounds) {
-  const [s, w, n, e] = bounds;
-  const bbox = `${s},${w},${n},${e}`;
-  const query = `
-[out:json][timeout:90];
-(
-  way["leisure"~"park|garden"]["name"](${bbox});
-  way["place"="square"]["name"](${bbox});
-  way["landuse"="recreation_ground"]["name"](${bbox});
-  rel["leisure"~"park|garden"]["name"](${bbox});
-);
-out geom tags;
-`.trim();
-
-  const data = await queryOverpass(query);
-  const results = [];
-
-  for (const el of data.elements ?? []) {
-    const tags = el.tags ?? {};
-    const name = tags.name;
-    if (!name) continue;
-
-    let coords = [];
-
-    if (el.type === 'way' && Array.isArray(el.geometry)) {
-      coords = el.geometry.map((p) => [p.lon, p.lat]);
-    } else if (el.type === 'relation' && Array.isArray(el.members)) {
-      for (const member of el.members) {
-        if (member.role === 'outer' && Array.isArray(member.geometry)) {
-          coords.push(...member.geometry.map((p) => [p.lon, p.lat]));
-        }
-      }
-      if (coords.length === 0) {
-        for (const member of el.members) {
-          if (Array.isArray(member.geometry)) {
-            coords.push(...member.geometry.map((p) => [p.lon, p.lat]));
-          }
-        }
-      }
-    }
-
-    if (coords.length < 3) continue;
-
-    let sumLng = 0, sumLat = 0;
-    for (const [lng, lat] of coords) { sumLng += lng; sumLat += lat; }
-    const centerLng = sumLng / coords.length;
-    const centerLat = sumLat / coords.length;
-
-    const lats = coords.map((c) => c[1]);
-    const lngs = coords.map((c) => c[0]);
-    const heightM = (Math.max(...lats) - Math.min(...lats)) * 111320;
-    const widthM = (Math.max(...lngs) - Math.min(...lngs)) * 111320 * Math.cos(centerLat * Math.PI / 180);
-    const extent = Math.max(widthM, heightM);
-
-    results.push({
-      name,
-      type: tags.leisure || tags.place || tags.landuse || 'park',
-      lat: centerLat,
-      lng: centerLng,
-      geometry: coords,
-      extent: Math.round(extent),
-      osmType: el.type,
-      osmId: el.id,
-      tags,
-    });
-  }
-
-  console.log(`[overpass] ${results.length} park areas (${results.filter((r) => r.osmType === 'relation').length} relations)`);
-  return results;
-}
