+++ tests/pipeline/bikepaths-yml-integrity.test.mjs
@@ -6,7 +6,7 @@
  * about Ottawa's cycling infrastructure.
  */
 import { describe, it, expect, beforeAll } from 'vitest';
-import { queryOverpass } from '../../scripts/pipeline/lib/overpass.mjs';
+import { queryOverpass } from '../../scripts/pipeline/lib/overpass.ts';
 import { buildBikepathsPipeline } from '../../scripts/pipeline/build-bikepaths.ts';
 import { loadCityAdapter } from '../../scripts/pipeline/lib/city-adapter.mjs';
